+++ src/main/java/com/nhnacademy/illuwa/review/entity/Comment.java
@@ -1,11 +1,9 @@
-package com.nhnacademy.illuwa.entity.review;
+package com.nhnacademy.illuwa.review.entity;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
 
 @Getter
 @Entity
