+++ src/main/java/com/nhnacademy/illuwa/review/entity/Review.java
@@ -1,14 +1,10 @@
-package com.nhnacademy.illuwa.entity.review;
+package com.nhnacademy.illuwa.review.entity;
 
-import com.nhnacademy.illuwa.entity.book.Book;
+import com.nhnacademy.illuwa.book.entity.Book;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
 
 import java.time.LocalDateTime;
